Changed alot of stuff in a long period of time.
-Made all orbits realistic.

-Tweaked shaders as always.

-Completely changed the planet asset system.

-Changed all jpg images into png.

-Completely changed the planet parameter system.

-Got skybox rendering back in (no daylight cycle yet).

-Added a dating system that seems to be slightly off on orbit
calculations. I'm guessing leap-years are the problem.

-Added a much easier to use shader system.

-Cleaned up just about the whole space gui rendering pipeline.

+++ src/main/java/com/team/futurecraft/FutureCraft.java
@@ -19,7 +19,7 @@ import net.minecraftforge.fml.relauncher.SideOnly;
 public class FutureCraft {
 	public static final String MODID = "futurecraft";
 	public static final String VERSION = "0.2";
-	public static float timeOffset = 0;
+	public static long timeOffset = 0;
 	
 	//creative tab
 	public static final CreativeTabs tabFutureCraft = new CreativeTabs("futurecraft") {
